Decouple stack from VM
